requirements.txt
fastapi
uvicorn
transformers
torch
python-dotenv
jinja2
pytest
pre-commit
python-multipart
